随机延迟

+++ MainWindow.h
@@ -30,7 +30,7 @@ public:
     const QList<BitMaskEditorDialog::Item>& getTabs() const { return tabs; }
     BitMaskEditorDialog::Item& getTab() { return tabs[tabWidget->currentIndex()]; }
     void showSupportDialog();
-    double getRandomDelay();
+    float getRandomDelay();
 
     void relayoutDevices();
